Updated heading, updated pom.xml, and fixed bugs.

+++ src/main/java/org/example/ex5/Numbers.java
@@ -1,11 +1,16 @@
+/*
+ *  UCF COP3330 Summer 2021 Assignment 1 Solutions
+ *  Copyright 2021 Crystal Gagne
+ */
+
 package org.example.ex5;
 
 import java.util.Scanner;
 
 public class Numbers {
-    public static Integer getNumber()
+    public static Double getNumber()
     {
         Scanner newNumber = new Scanner(System.in);
-        return Integer.parseInt(newNumber.nextLine());
+        return Double.parseDouble(newNumber.nextLine());
     }
 }
